requirements.txt
asgiref==3.2.7
astroid==2.4.1
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.4
django-environ==0.4.5
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-camel-case==1.1.2
djoser==2.0.3
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
plaid-python==3.7.0
pylint==2.5.2
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
